Namelist追加コンポーネント
